+++ CHANGELOG.md
@@ -8,6 +8,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Fixed
 - Improve Ledger support
 
+### Added
+- Trezor singlesig address display
+
 ## [3.8.5] - 2022-07-15
 ### Added
 - Enable watch-only in Liquid